feat: refactor PoolConfiguration component by modularizing sections into dedicated components for improved readability and maintainability
